Change order of function arguments for compatibility with 0.3 release...
This is primarily to avoid a situation where calling:

    gumbo.parse("input", 4)

...would result in the second argument being coerced to a string and
used as the value for GumboOptions::fragment_context. This would enable
fragment parsing mode and produce very different behaviour than the same
call running on lua-gumbo 0.3.

This is an API break during development, in order to retain API
compatibility across releases.