Reduce tree depth limit of parser from 800 to 400
During testing on OpenBSD 6.0, the test case with an input of 500 nested
div elements actually caused a stack overflow with Lua 5.3 (but not 5.2
or 5.1). Lowering the depth limit to 400 seems to solve the issue.

Allowing nesting of 800 levels deep was absurd anyway. Even WebKit's
default limit is only 512. Documents with more than a few hundred levels
of nesting are either badly authored, buggy or malicious.