Remove test case that asserted "gumbo.parseFile('.')" should fail
Trying to pass a directory name to the parseFile function fails on Linux
but successfully opens and reads (like an empty file) on OpenBSD 6.

What happens when trying to read from a directory thus seems to be
platform dependent, so there shouldn't be a test case that expects
a certain behaviour unless the behaviour is enforced in a portable
way within the parseFile function.