Fix duplicate TagSet initializer being ignored in is_special_node()
The TAG macros just expand to designated initializers for the array
index of the enum value of the named tag. Using any of these macros more
than once with the same tag name within a single TagSet just produces
duplicate initializers.

The value used is whichever initializer appears last in the list and any
prior initializers with the same index are effectively ignored. In this
case the intended effects of using "TAG(TITLE)" were being canceled out
by "TAG_SVG(TITLE)", which appeared later in the list.