Use "-text diff" git attributes for tree-construction test files
Unsetting the "text" attribute disables end-of-line normalization on
checkin and checkout. This is the desired behaviour for these files,
since they should be exact copies of the upstream html5lib tests.

The "diff" attribute is also explicitly set because git refuses to do
text diffs on two of the files, due to them containing test cases with
mixed line-endings (which it detects as "binary").