Remove vertical tab ("\v") from ASCII_SPACE category in lib/ascii.c
ISO C and POSIX both consider this a whitespace character, but
apparently the HTML and DOM specs don't. Since the use cases in
this codebase are going to be in the context of HTML, we should
follow the HTML definition. Although in practice it hardly makes a
difference, since vertical tab characters are practically obsolete.